Redaction_API/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
openai>=1.3.0 # Asegúrate de que sea la versión que soporta base_url
pydantic>=2.0.0
# Jinja2 no es estrictamente necesario con el enfoque actual, pero si lo usas:
# jinja2>=3.0.0